+++ django_prov/utils.py
@@ -3,13 +3,12 @@ import platform
 import sys
 
 
-def get_system_info_attributes(label):
+def get_system_info_attributes(label="sys"):
     """
     Returns a List of attributes customized for the ProvDocument.
     :param label: Prefix of the namespace that has to be taken
     :return: List of Attributes
     """
-    attributes = [(f'sys:python_version', sys.version)]
-    attributes.append((f'sys:os', platform.platform()))
-    attributes.append((f'sys:os_username', getpass.getuser()))
+    attributes = [(f'{label}:python_version', sys.version), (f'{label}:os', platform.platform()),
+                  (f'{label}:os_username', getpass.getuser())]
     return attributes